refactor: change to big serial

+++ internal/dto/document_client/get.go
@@ -1,11 +1,11 @@
 package dto
 
 type GetDocumentClientReq struct {
-	Id int32 `json:"id" validate:"required"`
+	Id int64 `json:"id" validate:"required"`
 }
 
 type GetDocumentClientRes struct {
-	Id         int32  `json:"id"`
+	Id         int64  `json:"id"`
 	ClientId   int32  `json:"clientId"`
 	DocumentId int32  `json:"documentId"`
 	Action     string `json:"action"`
